requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
requests
openpyxl
mysql.connector
discord.py